requirements.txt
tensorflow
scikit-learn
pandas
opencv-python
numpy
matplotlib
Pillow
flask
